+++ examples/batchedcalls/server.py
@@ -1,7 +1,7 @@
 from __future__ import print_function
 import time
 import Pyro4
-from Pyro4.socketutil import getMyIpAddress
+from Pyro4.socketutil import getIpAddress
 
 class Thingy(object):
     def multiply(self,a,b):
@@ -19,7 +19,7 @@ class Thingy(object):
         print(message)
         return 0
 
-d=Pyro4.Daemon(host=getMyIpAddress(workaround127=True), port=0)
+d=Pyro4.Daemon(host=getIpAddress("", workaround127=True), port=0)
 uri=d.register(Thingy(), "example.batched")
 print("server object uri:",uri)
 print("batched calls server running.")