+++ src/Pyro4/configuration.py
@@ -17,7 +17,7 @@ class Configuration(object):
     __slots__=("HOST", "NS_HOST", "NS_PORT", "NS_BCPORT", "NS_BCHOST",
                "COMPRESSION", "SERVERTYPE", "DOTTEDNAMES", "COMMTIMEOUT",
                "POLLTIMEOUT", "THREADING2", "ONEWAY_THREADED",
-               "DETAILED_TRACEBACK", "SOCK_REUSE",
+               "DETAILED_TRACEBACK", "SOCK_REUSE", "PREFER_IP_VERSION",
                "THREADPOOL_MINTHREADS", "THREADPOOL_MAXTHREADS",
                "THREADPOOL_IDLETIMEOUT", "HMAC_KEY", "AUTOPROXY",
                "BROADCAST_ADDRS", "NATHOST", "NATPORT", "MAX_MESSAGE_SIZE",
@@ -55,6 +55,7 @@ class Configuration(object):
         self.MAX_MESSAGE_SIZE = 0   # 0 = unlimited
         self.BROADCAST_ADDRS = "<broadcast>, 0.0.0.0"   # comma separated list of broadcast addresses
         self.FLAME_ENABLED = False
+        self.PREFER_IP_VERSION = 4    # 4, 6 or 0 (let OS choose according to RFC 3484)
 
         if useenvironment:
             # process enviroment variables